Move subscription-related inner classes to the appropriate classes
